Add Day1 solution and input data file
This commit consists of the addition of a Java file, Day1.java, which contains a solution to a problem. The problem is about finding the sum of the first and last digit from every line in a given file. The solution uses Files API to read 'input.txt' line by line and a